Change PushInterval type to int64

+++ types/types.go
@@ -3,7 +3,7 @@ package types
 type SensorStatus struct {
 	SensorGroup   string `json:"sensorGroup,omitempty" form:"sensorGroup"`
 	SensorAddress string `json:"sensorAddress,omitempty" form:"sensorAddress"`
-	PushInterval  int    `json:"pushInterval,omitempty" form:"pushInterval"`
+	PushInterval  int64  `json:"pushInterval,omitempty" form:"pushInterval"`
 	InfluxHost    string `json:"influxHost,omitempty" form:"influxHost"`
 	InfluxPort    string `json:"influxPort,omitempty" form:"influxPort"`
 	InfluxOrg     string `json:"influxOrg,omitempty" form:"influxOrg"`
